refactor(tests): split staleness gate termination tests into their own module
Round-5 coverage pushed test_check_generated_staleness.py to 509 lines,
past the 500-line test file-size ceiling the taste ratchet enforces (the
+1 regression was this file). The termination and budget tests move to
test_check_generated_staleness_termination.py; the fake-repo stubs and
the clamp-env hygiene fixture move to staleness_gate_helpers.py so both
modules share one copy. 34 tests pass; taste ratchet back at baseline
583.

Refs #5079
