+++ tests/validation/test_check_generated_staleness.py
@@ -13,6 +13,10 @@ cannot see whether any caller invokes it, which is the shape that let
 below drives the real sequence and spies on the rebindable module attribute
 that ``_root_only`` resolves at call time.
 
+Deadline, budget, and termination behavior lives in
+``test_check_generated_staleness_termination.py``; shared stubs live in
+``staleness_gate_helpers.py``.
+
 Coverage:
 
 - positive: both checks clean gives exit 0 and prints the examined count.
@@ -20,7 +24,7 @@ Coverage:
   a non-directory root gives exit 2; a killed child gives external exit 3;
   removing the gate from ``_SEQUENCE`` fails the wiring test.
 - edge: a failing first check leaves the second unrun, asserted on a marker
-  file the second stub would have written.
+  file the second stub would have written; the failure echo is tail-bounded.
 """
 
 from __future__ import annotations
@@ -28,7 +32,6 @@ from __future__ import annotations
 import io
 import subprocess
 import sys
-import time
 from collections.abc import Callable
 from contextlib import redirect_stdout
 from pathlib import Path
@@ -36,60 +39,40 @@ from types import SimpleNamespace
 
 import pytest
 
-REPO_ROOT = Path(__file__).resolve().parents[2]
+from tests.validation.staleness_gate_helpers import (
+    REPO_ROOT,
+    build_all_ran,
+    check_generated_staleness,
+    fake_repo,
+    no_ambient_outer_cap,  # noqa: F401  (autouse fixture)
+    stub,
+)
 
-# Import the way production imports (issue #2223): prepend ``scripts/validation``
-# to ``sys.path`` and import by bare name.
+# Import the way production imports (issue #2223): prepend
+# ``scripts/validation`` to ``sys.path`` (idempotent; the helpers module does
+# it too) and import by bare name. The statement between the import blocks is
+# load-bearing: it stops import sorting from hoisting ``pre_pr_sequence``
+# above the path setup it depends on.
 _VALIDATION_DIR = REPO_ROOT / "scripts" / "validation"
 if str(_VALIDATION_DIR) not in sys.path:
     sys.path.insert(0, str(_VALIDATION_DIR))
-import check_generated_staleness
 import pre_pr_sequence
 
 GATE_NAME = "Generated Artifact Staleness"
-
-
-@pytest.fixture(autouse=True)
-def _no_ambient_outer_cap(monkeypatch: pytest.MonkeyPatch) -> None:
-    """A developer shell exporting the clamp variable must not skew tests
-    that reason about the unclamped aggregate budget."""
-    monkeypatch.delenv(check_generated_staleness._OUTER_CAP_ENV, raising=False)
-
-
-def _stub(path: Path, body: str) -> None:
-    """Write an executable-by-interpreter stub at ``path``."""
-    path.parent.mkdir(parents=True, exist_ok=True)
-    path.write_text(f"import sys\n{body}\n", encoding="utf-8")
-
-
-def _fake_repo(tmp_path: Path, sync_exit: int, build_exit: int) -> Path:
-    """A root holding stubs at the two real script paths the gate invokes."""
-    _stub(tmp_path / "scripts" / "sync_plugin_lib.py", f"sys.exit({sync_exit})")
-    _stub(
-        tmp_path / "build" / "scripts" / "build_all.py",
-        "from pathlib import Path\n"
-        "Path(__file__).with_name('build_all_ran.marker').write_text('1')\n"
-        f"sys.exit({build_exit})",
-    )
-    return tmp_path
-
-
-def _build_all_ran(repo_root: Path) -> bool:
-    return (repo_root / "build" / "scripts" / "build_all_ran.marker").is_file()
 
 
 class TestExitCodes:
     """A detected violation must reach the caller as a non-zero exit."""
 
     def test_a_clean_tree_exits_zero(self, tmp_path: Path) -> None:
-        root = _fake_repo(tmp_path, sync_exit=0, build_exit=0)
+        root = fake_repo(tmp_path, sync_exit=0, build_exit=0)
 
         assert check_generated_staleness.main([str(root)]) == 0
 
     def test_a_clean_run_reports_the_examined_count(
         self, tmp_path: Path, capsys: pytest.CaptureFixture[str]
     ) -> None:
-        root = _fake_repo(tmp_path, sync_exit=0, build_exit=0)
+        root = fake_repo(tmp_path, sync_exit=0, build_exit=0)
 
         check_generated_staleness.main([str(root)])
 
@@ -98,12 +81,12 @@ class TestExitCodes:
     def test_build_all_staleness_exits_one(self, tmp_path: Path) -> None:
         # build_all.py --check exits 2 on staleness; the gate maps every
         # non-zero child exit onto ADR-035 exit 1.
-        root = _fake_repo(tmp_path, sync_exit=0, build_exit=2)
+        root = fake_repo(tmp_path, sync_exit=0, build_exit=2)
 
         assert check_generated_staleness.main([str(root)]) == 1
 
     def test_sync_drift_exits_one(self, tmp_path: Path) -> None:
-        root = _fake_repo(tmp_path, sync_exit=1, build_exit=0)
+        root = fake_repo(tmp_path, sync_exit=1, build_exit=0)
 
         assert check_generated_staleness.main([str(root)]) == 1
 
@@ -113,7 +96,7 @@ class TestExitCodes:
         # The module's exit table promises 2 for an absent script: the gate
         # could not run and the script needs restoring, where the drift remedy
         # (regenerate and commit) would be the wrong action.
-        root = _fake_repo(tmp_path, sync_exit=0, build_exit=0)
+        root = fake_repo(tmp_path, sync_exit=0, build_exit=0)
         (root / "build" / "scripts" / "build_all.py").unlink()
 
         assert check_generated_staleness.main([str(root)]) == 2
@@ -130,8 +113,8 @@ class TestExitCodes:
     ) -> None:
         # ADR-035: a timeout kill means the tree was never scored, which is
         # external (3), not drift (1). The stub sleeps past a 1s cap.
-        root = _fake_repo(tmp_path, sync_exit=0, build_exit=0)
-        _stub(
+        root = fake_repo(tmp_path, sync_exit=0, build_exit=0)
+        stub(
             root / "scripts" / "sync_plugin_lib.py",
             "import time\ntime.sleep(30)\nsys.exit(0)",
         )
@@ -145,7 +128,7 @@ class TestExitCodes:
         # The timeout branch must preserve TimeoutExpired.stdout/.stderr and
         # append the kill marker, not discard the diagnosis the child already
         # emitted.
-        root = _fake_repo(tmp_path, sync_exit=0, build_exit=0)
+        root = fake_repo(tmp_path, sync_exit=0, build_exit=0)
 
         class FakeProc:
             returncode = None
@@ -186,213 +169,6 @@ class TestExitCodes:
         assert "exceeded 1.0s" in output
 
 
-class TestBoundedGracefulTermination:
-    """The gate is bounded as a whole, and expiry preserves child cleanup."""
-
-    def test_the_gate_budget_is_positive_and_shared(self) -> None:
-        # An unbounded child stalls pre_pr.py indefinitely for callers not
-        # under the lefthook job cap (reliability review on PR #5088). The
-        # budget is one aggregate so the worst case is the budget, not the
-        # budget times the row count.
-        assert check_generated_staleness._GATE_BUDGET_SECONDS > 0
-
-    def test_budget_plus_grace_fits_inside_the_lefthook_cap(self) -> None:
-        # lefthook kills the whole pre-pr-validation process tree at its cap,
-        # without the SIGINT path, so the gate's worst case (budget + grace)
-        # must leave the rest of the sequence room inside that cap. Parsed
-        # from the live lefthook.yml so a cap change breaks this loudly.
-        import yaml
-
-        config = yaml.safe_load((REPO_ROOT / "lefthook.yml").read_text())
-
-        def _jobs(node: object) -> list[dict]:
-            found: list[dict] = []
-            if isinstance(node, dict):
-                if "jobs" in node:
-                    for item in node["jobs"]:
-                        found.extend(_jobs(item))
-                if isinstance(node.get("group"), dict):
-                    found.extend(_jobs(node["group"]))
-                if "name" in node:
-                    found.append(node)
-            return found
-
-        jobs = _jobs(config.get("pre-push", {}))
-        caps = [j for j in jobs if j.get("name") == "pre-pr-validation"]
-        assert caps, "pre-pr-validation job not found in lefthook.yml"
-        timeout = caps[0]["timeout"]
-        assert isinstance(timeout, str) and timeout.endswith("m")
-        cap_seconds = int(timeout[:-1]) * 60
-
-        worst_case = (
-            check_generated_staleness._GATE_BUDGET_SECONDS
-            + check_generated_staleness._TERMINATION_GRACE_SECONDS
-        )
-        assert worst_case <= cap_seconds / 2, (
-            f"gate worst case {worst_case}s must leave at least half of the"
-            f" {cap_seconds}s pre-pr-validation cap for the rest of the"
-            f" sequence"
-        )
-
-        # The job declares its own cap to the gate so the runtime clamp can
-        # key off it (review round 5). The declaration must equal the job's
-        # actual timeout, or the clamp bounds against a fiction.
-        declared = caps[0].get("env", {}).get(
-            check_generated_staleness._OUTER_CAP_ENV
-        )
-        assert declared is not None, (
-            f"pre-pr-validation must declare"
-            f" {check_generated_staleness._OUTER_CAP_ENV}"
-        )
-        assert float(declared) == cap_seconds
-
-    def test_an_exhausted_budget_reports_external_without_running_the_child(
-        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        root = _fake_repo(tmp_path, sync_exit=0, build_exit=0)
-        monkeypatch.setattr(check_generated_staleness, "_GATE_BUDGET_SECONDS", -1.0)
-
-        assert check_generated_staleness.main([str(root)]) == 3
-        assert not _build_all_ran(root)
-
-    def test_expiry_lets_the_child_finally_run(self, tmp_path: Path) -> None:
-        # The whole point of graceful termination: SIGKILL would skip the
-        # child's finally (which is what restores build_all's snapshot), but
-        # SIGINT raises KeyboardInterrupt so the finally runs. The stub
-        # mirrors that shape: it sleeps past the deadline and writes a marker
-        # from its finally. The marker existing proves cleanup ran.
-        if sys.platform == "win32":
-            pytest.skip("POSIX SIGINT path; non-POSIX uses terminate()")
-        child = tmp_path / "cleanup_child.py"
-        marker = tmp_path / "cleanup_ran.marker"
-        child.write_text(
-            "import sys, time\n"
-            "try:\n"
-            "    time.sleep(30)\n"
-            "finally:\n"
-            f"    open({str(marker)!r}, 'w').write('1')\n",
-            encoding="utf-8",
-        )
-
-        code, output = check_generated_staleness._run_check(child, tmp_path, 1.0)
-
-        assert code is None
-        assert marker.is_file(), "the child's finally never ran"
-        assert "exceeded 1.0s" in output
-        assert "honored the interrupt" in output
-
-    def test_a_child_that_ignores_the_interrupt_is_killed_with_a_warning(
-        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        # The escalation branch: a child that shrugs off SIGINT must still be
-        # bounded, and the caller must be told the tree may hold partial
-        # writes, because that is the one path where cleanup did not run.
-        if sys.platform == "win32":
-            pytest.skip("POSIX SIGINT path; non-POSIX uses terminate()")
-        child = tmp_path / "stubborn_child.py"
-        child.write_text(
-            "import signal, sys, time\n"
-            "signal.signal(signal.SIGINT, signal.SIG_IGN)\n"
-            "print('armored', flush=True)\n"
-            "time.sleep(60)\n",
-            encoding="utf-8",
-        )
-        monkeypatch.setattr(
-            check_generated_staleness, "_TERMINATION_GRACE_SECONDS", 1.0
-        )
-
-        code, output = check_generated_staleness._run_check(child, tmp_path, 2.0)
-
-        assert code is None
-        assert "ignored the interrupt and was killed" in output
-        assert "partial generated writes" in output
-
-    def test_a_non_posix_host_terminates_instead_of_signaling(
-        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        # Windows has no SIGINT delivery to a child; the fallback is
-        # terminate(). This drives the branch with a fake so the suite pins
-        # it on every platform: on expiry the non-POSIX path must call
-        # terminate() and must not attempt a POSIX signal.
-        class FakeProc:
-            returncode = None
-
-            def __init__(self) -> None:
-                self.calls = 0
-                self.terminated = False
-                self.signals: list[int] = []
-
-            def communicate(self, timeout: float | None = None) -> tuple[str, str]:
-                self.calls += 1
-                if self.calls == 1:
-                    raise subprocess.TimeoutExpired(cmd=["x"], timeout=timeout or 0.0)
-                return ("", "")
-
-            def send_signal(self, sig: int) -> None:
-                self.signals.append(sig)
-
-            def terminate(self) -> None:
-                self.terminated = True
-
-            def kill(self) -> None:
-                return None
-
-        fake = FakeProc()
-        monkeypatch.setattr(
-            check_generated_staleness.subprocess, "Popen", lambda *a, **k: fake
-        )
-        monkeypatch.setattr(check_generated_staleness.os, "name", "nt")
-
-        code, _output = check_generated_staleness._run_check(
-            tmp_path / "any.py", tmp_path, 1.0
-        )
-
-        assert code is None
-        assert fake.terminated, "non-POSIX expiry must call terminate()"
-        assert fake.signals == [], "non-POSIX expiry must not send POSIX signals"
-
-    def test_a_spent_outer_share_reports_external_without_running_the_child(
-        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        # The round-5 clamp: the gate's clock starts when the sequence
-        # reaches it, after earlier validations spent part of the same outer
-        # lefthook timer. When the declared cap leaves less than the grace
-        # window, spawning a child would invite the outer SIGKILL mid-write,
-        # so the gate must refuse to spawn at all.
-        root = _fake_repo(tmp_path, sync_exit=0, build_exit=0)
-        monkeypatch.setenv(check_generated_staleness._OUTER_CAP_ENV, "1")
-
-        assert check_generated_staleness.main([str(root)]) == 3
-        assert not _build_all_ran(root)
-
-    def test_an_unspent_outer_cap_leaves_the_gate_running(
-        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        # Control for the clamp test: a freshly started process under the
-        # real declared cap must behave exactly as with no clamp at all.
-        root = _fake_repo(tmp_path, sync_exit=0, build_exit=0)
-        monkeypatch.setenv(check_generated_staleness._OUTER_CAP_ENV, "900")
-        monkeypatch.setattr(
-            check_generated_staleness, "_PROCESS_START", time.monotonic()
-        )
-
-        assert check_generated_staleness.main([str(root)]) == 0
-        assert _build_all_ran(root)
-
-    def test_a_malformed_outer_cap_warns_and_disables_the_clamp(
-        self, tmp_path: Path, monkeypatch: pytest.MonkeyPatch,
-        capsys: pytest.CaptureFixture[str],
-    ) -> None:
-        # A typo in the hook configuration must not block every contributor's
-        # push: the clamp only tightens an already-bounded gate, so the
-        # failure mode is a loud warning plus the unclamped budget.
-        root = _fake_repo(tmp_path, sync_exit=0, build_exit=0)
-        monkeypatch.setenv(check_generated_staleness._OUTER_CAP_ENV, "banana")
-
-        assert check_generated_staleness.main([str(root)]) == 0
-        assert "[WARN]" in capsys.readouterr().err
-
-
 class TestGeneratorOrder:
     """sync before build, per .claude/rules/generated-artifacts.md."""
 
@@ -405,19 +181,19 @@ class TestGeneratorOrder:
         # The isolating assertion is the point: pytest would pass on the exit
         # code alone even if build_all ran first and its verdict was compared
         # against a stale .claude/lib.
-        root = _fake_repo(tmp_path, sync_exit=1, build_exit=0)
+        root = fake_repo(tmp_path, sync_exit=1, build_exit=0)
 
         assert check_generated_staleness.main([str(root)]) == 1
-        assert not _build_all_ran(root)
+        assert not build_all_ran(root)
 
     def test_the_control_run_does_reach_build_all(self, tmp_path: Path) -> None:
         # Without this control, the assertion above would pass against a gate
         # that never invokes build_all at all.
-        root = _fake_repo(tmp_path, sync_exit=0, build_exit=0)
+        root = fake_repo(tmp_path, sync_exit=0, build_exit=0)
 
         check_generated_staleness.main([str(root)])
 
-        assert _build_all_ran(root)
+        assert build_all_ran(root)
 
 
 class TestEchoTail:
